feat: login 구현

+++ src/main/java/edu/allinone/sugang/repository/StudentRepository.java
@@ -5,7 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface StudentRepository extends JpaRepository<Student, Long> {
+public interface StudentRepository extends JpaRepository<Student, Integer> {
 
+    // student_number 필드를 기준으로 'Student' 엔티티 조회하는 메서드
     Optional<Student> findByStudentNumber(String studentNumber);
 }
